Move make_stitched_labels from experimental.im to experimental.tl
Co-locate all four steps of the tiling-QC / stitching workflow
(calculate_tiling_qc, assign_stitch_groups, make_stitched_labels, and
the pl.tiling_qc viz) so the pipeline is discoverable in one place.
Previously make_stitched_labels lived in `im` (it emits a labels
raster), splitting the workflow across `tl` and `im`.

Pure move: the module uses absolute imports, so no internal import
changes are needed. Updates both __init__ exports, docs/api.md, and
the test references.
