+++ docs/api.md
@@ -152,7 +152,7 @@ See the {doc}`extensibility guide </extensibility>` for how to implement a custo
     experimental.tl.TilingQCParams
     experimental.tl.assign_stitch_groups
     experimental.tl.StitchParams
-    experimental.im.make_stitched_labels
+    experimental.tl.make_stitched_labels
     experimental.pl.tiling_qc
     experimental.im.fit_stain_reference
     experimental.im.normalize_stains
