+++ src/squidpy/experimental/tl/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
+from ._stitched_labels import make_stitched_labels
 from ._tiling_qc import TilingQCParams, calculate_tiling_qc
 from ._tiling_stitch import StitchParams, assign_stitch_groups
 
-__all__ = ["StitchParams", "TilingQCParams", "assign_stitch_groups", "calculate_tiling_qc"]
+__all__ = ["StitchParams", "TilingQCParams", "assign_stitch_groups", "calculate_tiling_qc", "make_stitched_labels"]
