+++ src/squidpy/experimental/tl/_stitched_labels.py
